Invent openqa_onebox_test group to try and set up one-box openQA
This is an attempt to add a group that'll set openqa-x86-worker05
as a one-box openQA instance which doesn't report results. This
is to try and help with debugging
https://bugzilla.redhat.com/show_bug.cgi?id=2009585 : the idea is
to have a disposable openQA instance where we don't care if all
the tests fail, so we can try out various things to resolve that
bug on it. Debugging on lab is awkward because we do actually
care about lab's results, especially since it's the only instance
testing Rawhide updates.

This requires quite a lot of surgery to get around various
assumptions in the existing groups and plays (no reporting of
results, no NFS...) and make the box its own postgresql server.

Signed-off-by: Adam Williamson <awilliam@redhat.com>