bodhi / backend: run owner-sync-pagure as apache user
This has been broken since basically forever.
The bodhi keytab is owner by apache and expects to be run under that
user. Two reasons we never noticed/fixed this:

Until recently it was failing with a error trying to talk to pagure.io's
api. So I just thought it was too slow and needed to be optomized.
When I tweaked pagure's httpd config for the DDoS we had the other day I
fixed this on the server end, then it got to the auth error.

Signed-off-by: Kevin Fenzi <kevin@scrye.com>