Add nirik's nbde client fixups to openQA worker playbook
nirik did some stuff in the virthost and buildhw groups to try
and deal with a problem where network configuration created
during the initrd phase for the nbde_client role is activated
by NM during the regular system boot phase, which results in
the network configuration not being the one we actually want
and carefully set up. However, he didn't add this stuff to the
openqa-workers group playbook even though that uses the
nbde_client role too, and we sure are having the same problem
on the openQA workers. Adding it now to see if it helps.

Signed-off-by: Adam Williamson <awilliam@redhat.com>