FIX: nagios external was referencing phx2 ip addresses
The PHX2 colocation has been turned off. This meant that some configs
which had been accidently working before due to referencing an ip
address there that no longer existed broke. The fix was to rewrite the
config so that it contained proper router ips and remove all mentions
of the PHX2 ip address.

Signed-off-by: Stephen Smoogen <ssmoogen@redhat.com>