bodhi: when sending to greenwave, always use verbose
This is a hack to work around bugs in bodhi and greewave.
It happens if we send here with verbose, bodhi sends the right stuff and
avoids the greenwave bug. This should be dropped as soon as upstream
fixes land in bodhi or greenwave. See:
https://pagure.io/greenwave/pull-request/666
and
https://github.com/fedora-infra/bodhi/pull/4687
for the proper fix.

Signed-off-by: Kevin Fenzi <kevin@scrye.com>