bkernel: add config for nspawn bind mount
We want to keep the old mock bind mount for non rawhide branches, but
rawhide is using nspawn, so we want to add a directive there to pass
'--bind' to it to correctly mount the pesign socket directory so kernels
can be signed for secure boot.

See https://github.com/rpm-software-management/mock/issues/140

Moving forward this could be fixed in mock, in which case we remove the
nspawn args. Or it could be fixed by pesign moving the socket directory,
in which case we remove nspawn args and adjust the old mock bind mount
to the new location. For now, this works around the current crop of
issues.

Signed-off-by: Kevin Fenzi <kevin@scrye.com>