[release-monitoring] Generate docs in virtual environment
I'm not sure what is happening, but the docs and bootstrap symlink are not
installed in the system. Let's try to first install Anitya to virtual
environment and build the docs and after that do a direct install to container.

Also don't remove the cloned folder, so we can look what is happening there.

Signed-off-by: Michal Konečný <mkonecny@redhat.com>