osbs: fix race between stg and prod making registry secrets
If you run these playbooks without any limit, The 4 master hosts (2
x86_64 and 2 aarch64) will be in the play, but they all use local_action
to make the local secrets file thats loaded. This means, whichever of
them happens to be writing the file last, thats the version of the file
that all 4 of them get. This is particularly bad when it's the staging
creds and the prod hosts get it loaded. :(

So, adding {{ env }} here makes the staging and prod versions seperate
so they don't step on each other.

Signed-off-by: Kevin Fenzi <kevin@scrye.com>