greenwave: split the openQA update test policies by group
This kinda goes along with:
https://github.com/fedora-infra/bodhi/pull/4759
but neatly, we don't need to wait for that to be merged or
deployed, we can do it right away and it'll work fine now but
also DTRT when we switch to grouped mode in Bodhi. This splits the
greenwave policies for openQA update tests into groups according
to the test flavors, and has each grouped policy apply to what
will be the correct decision contexts for Bodhi's grouped critpath
support. We *also* have each grouped policy apply to the existing
non-grouped contexts, so as long as Bodhi is running in
non-grouped mode, things should behave exactly the same: all
critpath updates will be gated on all the tests.

We ensure every critpath group is listed in the null policies,
to make sure we don't get errors when Bodhi queries against them.

Signed-off-by: Adam Williamson <awilliam@redhat.com>