Adjust running koji-split cleanup to 1/week from 1/month
The reposplit cron job runs on the first day of the month and deletes
anything older than 30 days old at that time. However this has caused
a disk crunch on centos stream as each day is 195 GB of data and 60
days of traffic was over 8TB. The cron job is now to run every Sunday
and remove any data older than 14 days.

The same is done for the RHEL splitter as its space was growing also.

Signed-off-by: Stephen Smoogen <ssmoogen@redhat.com>