Add additional scripts
Both scripts help when cleaning up badges. All authorizations and
all invitations as well as awardees need to be removed before the
badge itself can be removed.

There's no functionality in the web interface to see who's
authorized for particular badges.

Invitations are a bit of an undocumented feature. But I stumbled
upon it trying to delete an old badge.