openshift-apps: remove build step from a few apps
Everytime we run the playbook a new build kicks off, but
the app was just restarted. So what happens is we end up
with the app getting started twice (once when the
deploymentconfig gets updated and once when the build finishes).
This could be bad if the app has some startup steps that need
to not be interrupted.

Let's just manually trigger builds since we have the permissions
to do that in the web interface and via the CLI.