buildvm_s390x: attempt to rebalance kvm lpar to allow for heavybuilders
We have been having issues with webkitgtk not being able to build due to
memory constraints on the existing builders. Also, we are overcomitted
on memory on the kvm lpar. So, to hopefully fix this:

* remove 3 existing builders.
* just leave the 3 cpus and 17gb memory from one free for the host
* make 2 of the other builders double the size in memory, cpu and disk.
* Will add these 2 to the heavybuilder channel and hopefully webkitgtk
  will be happy again.

I'm a bit concerned that this might slow the mass rebuild down, but we
will see. :)

Signed-off-by: Kevin Fenzi <kevin@scrye.com>