cloud-noc-os01: add group and access to sysadmin-copr
This noc host in rdu-cc is used for pxe installs, dhcp server and also
to access the private mgmt network for hosts in rdu-cc. Setup it's own
group and add sysadmin-copr to access it so they can get to the mgmt
interfaces on various copr machines there to reboot them.

Signed-off-by: Kevin Fenzi <kevin@scrye.com>