releng-bot: drop alias in favor of fixing address in fas
Right now releng-bot has a fas address of 'releng-bot@fedoraproject.org'
which is... confusing. The alias overrides this and sends email to
admin, but it results in a duplicate, causing the cron job to send mail
about the duplicate everytime newaliases run.

So, instead drop the alias here and switch the user in fas to be
admin+relengbot. This will still go to admin, not run into problems with
the address already in use in fas and should cause the newaliases to
stop complaining.

Signed-off-by: Kevin Fenzi <kevin@scrye.com>