defaults: set network_allow_restart: yes
linux system roles does a fine job configuring networking on our
systems, but without this it just configures it but doesn't bring things
'live' until a 'nmcli c up eth0'. Just set this so it should allow it to
restart things and reflect the network as we want it right after the
playbook runs on it.

Signed-off-by: Kevin Fenzi <kevin@scrye.com>