openshift / fedocal / staging: disable cron job in staging
The fedocal cron jobs fail in staging because they try to send to
'localhost' for smtp server. We could redirect them to use bastion, but
then people would get a bunch of reminders from prod and staging and get
confused by it. Ideally, fedocal would have a way to just print emails
to stdout instead of sending to SMTP server to use for testing them in
staging, but for now, until we have that just disable the cron job in
staging.

Signed-off-by: Kevin Fenzi <kevin@scrye.com>