bodhi: enable automated pushes all the time now
This converts the automated pushes cron template to just a simple file
thats always enabled. Since new bodhi has a way to 'freeze' frozen
releases, it won't push stable updates for any of those releases (but
will push the rest). So, we can just leave it always enabled as long as
we properly freeze/unfreeze releases.

Since it's just a simple file now, move it to one and change it from a
template.

Signed-off-by: Kevin Fenzi <kevin@scrye.com>