aliases: Update fcaic@ to fca@ to represent role title change
This commit makes a mostly cosmetic change for the Fedora Community
Architect email alias. In January 2023, the role title was changed from
Fedora Community Action & Impact Coordinator (FCAIC) to Fedora Community
Architect (FCA). There are additional communications planned for this
title change, but for now, this commit updates the mail aliases
accordingly for future use.

Additionally, the fcaic@ alias is preserved but it points to the fca@
email alias.

Furthermore, I also added the Fedora Project Leader @mattdm as a
recipient to the sponsors@ email alias, as we gear up for Flock again!

Signed-off-by: Justin W. Flory (he/him) <jwf@redhat.com>