openQA worker hosts: switch 03 and 06, cut workers on 05
openqa-x86-worker03 seems to be a bit poorly lately, it quite
often fails jobs in 'hardware blip' looking ways, even after a
reboot. It's also the equal-worst hardware in the worker host
pool with 05. So let's swap 03 and 06 so prod has most of the
best hardware, and lab has the poorly box. Also while doing a
quick hardware survey I noticed 05 is equally as underpowered
as 03 (it has 2x E5-2680v3, total 24 physical CPUs, all the
other hosts aside from those two have 2x16 core CPUs), so this
cuts its worker count to the same as 03 (and makes the comment
more accurate for both). Added comments to inventory with the
CPU info for each box for future reference.

Signed-off-by: Adam Williamson <awilliam@redhat.com>