download: add mirror.fcix.net to tier1 mirrors
fcix.net is a network of small mirrors that cache/serve popular content
from a number of free software projects. They mirror fedora and would
like to have access to tier0 mirrors. This adds them with their hostname
(so they can adjust the ips later if needed).

Thanks for mirroring!

Signed-off-by: Kevin Fenzi <kevin@scrye.com>