fasjson: try and get it to use dns to lookup kdc
We have been hard coding the kdc into the config here, but if we drop
that it should just look up the kdc from dns. This should also allow it
to use ipa02 and ipa03 instead of just fixating on 01. Hopefully that
will spread some load around and prevent timeouts we have been seeing.

Signed-off-by: Kevin Fenzi <kevin@scrye.com>