bodhi: adjust arches for updates pushes
Right now we are pushing a pointless armhfp variant in f37/f38 updates.
(It only has noarch packages in it).

So, lets clean up the logic here for both rpm updates pushes and modular
pushes. We can drop atomic as we don't do that anymore.

After we push this out and confirm it's working for a day or two, we
should go in and remove the bogus armhfp dirs from the master mirrors.

Signed-off-by: Kevin Fenzi <kevin@scrye.com>