[release-monitoring] Change the command syntax
When debugging the issue I tried to use `sleep 1000` as command for one of the
mid hook containers and this worked. So it seems that the only issue is with how
the command is actually written. So let's try a way that is in the OpenShift
documentation and see if this works.

Signed-off-by: Michal Konečný <mkonecny@redhat.com>