fix delete_old_oci_images.py
Previously, the delete_old_oci_images.py script sometimes caused a
traceback when trying to get the date from the JSON response for an OIC
image blob. This was because it was trying to read a JSON value that
didnt exist in some cases.

This change simply skips those images -- might update this in the future
to fins a fallback date to use, but for now just skip them.

Signed-off-by: Ryan Lerch <rlerch@redhat.com>