openQA aarch64: add new box to prod, shift others around
Thanks to nirik we have a new super-powerful openQA aarch64
worker host. For now, deploy it in production as the tap1 worker
and keep one old busted emag on prod as the tap2 worker. Move
the other old busted emag to lab so it has more capacity at
least, and give it a conventional tap1/tap2 setup. I might
tweak this again to have *only* the powerful worker in prod and
make it run all the tap jobs, to see if that's more reliable
than keeping an emag around. Or maybe I'll try and fiddle with
the classes so only the powerful box runs graphical tests, or
something.

Signed-off-by: Adam Williamson <awilliam@redhat.com>