fasjson: Enable flock-staff and flock-coc aliases again
With Flock around the corner, we need some email addresses for people to
reach the organizer team or for setting a reply-to alias for the new CFP
system we are introducing this year. All Flock aliases were disabled per
request of @mattdm in #10411 because of spam. It seems `flockinfo@` is
the problematic mail alias, so I took a conservative approach by only
enabling the `flock-staff@` and `flock-coc@` addresses.

I also added @shaunm to both aliases because he is also a part of the
core staff team for this year's Flock.

Signed-off-by: Justin W. Flory (he/him) <jwf@redhat.com>