[wiki] Remove old patched openIDConnect, and comment about new setup
We previously moved the FPCA check from a patched file held here in
ansible into the RPM itself when it was built for the Fedora infra
repos. While we no longer installed the patched file, it remained here
in the files dir of the wiki role, so we are deleting it.

Also, have added a comment to the play to say where we are holding this
patch now, so when i forget about it again, and come back to it in
another year or so, i can remember faster

Signed-off-by: Ryan Lerch <rlerch@redhat.com>