Retire old rdu2 space.
This space was in a networking cage thats going away and all the
machines there are old and out of warentee.
This space was intended for disaster recovery purposes.
For that now we have a internal vm that can access our mirrored netapp
storage, so we can sync anything off it we need to when iad2 is down.
I will be resetting up a batcave13 in another datacenter to allow us a
backup ansible/dns control host. bastion13/ns13/proxy13 will likely just
go away forever. download-rdu01 should be replaced by the new
download-cc-rdu01 once we have the new hardware in place for that.

Signed-off-by: Kevin Fenzi <kevin@scrye.com>