badges(tahrir): Add admins (@riecatnor, @smeragoel, @sumantrom)
This commit adds three new admins to the Tahrir web app, AKA the front-
end for `badges.fp.o`. This satisfies a request from @sumantrom to get
more involved with the management of the Badges site. I also added
@riecatnor and @smeragoel proactively due to their role as Fedora Badges
designers and co-mentors with the two Outreachy interns coming on board.

Partially closes fedora-badges#877:

https://pagure.io/fedora-badges/issue/877

Signed-off-by: Justin W. Flory (he/him) <jwf@redhat.com>