bugz.fedoraproject.org: replace %20 with + to fix redirect
The backslash-escaped space in 'Fedora EPEL' was replaced with %20 in
6e982da0d4 (bugz.fedoraproject.org: urlescape redirect instead of \,
2023-03-15).  Unfortunately, mod_rewrite treats the %2 as a
backreference, resulting in the string being 'Fedora0EPEL' which yeilds
no results for EPEL bugs.

We could possibly use on of the many RewriteRule flags to adjust the
handling of backreferences, but using a `+` seems like the simpler
option and works just as well in testing.

Signed-off-by: Todd Zullinger <tmz@pobox.com>