ipa: exclude the ipa postgres user in favor of local one
We have a postgres user in ipa to prevent people from making an account
with that name, but we need/want to use a local version of this on
database servers, not the ipa one. We need the local one because the ipa
one is locked and this prevents database backups from working.
(Locked accounts can't run cron jobs).

Signed-off-by: Kevin Fenzi <kevin@scrye.com>