aliases: Add new organizer to flock-staff alias, change flock-coc
This commit adds Natalie Pazmiño as a contact on the `flock-staff` email
alias. She is supporting from Red Hat in the global events team.
Additionally, I also changed the `flock-coc` alias to use `flock-staff`
instead of a smaller subset of people. If someone is emailing this
alias, it might be because of an emergency and more eyes is better than
fewer. If there is a less severe incident, this could carry through our
normal channels for Code of Conduct incidents.

Signed-off-by: Justin W. Flory (he/him) <jwf@redhat.com>