Temporary fix website by restoring multiview
It was thought that MultiViews was not needed for various websites to work and so was removed. It turns out that was wrong, and it was very much relied on for permissions of pages to work on the proxies and such. Mea-culpa mea-culpa

This fix restores the item until we are able to come up with a better fix.

Signed-off-by: Stephen Smoogen <ssmoogen@redhat.com>