proxies: redirect fas.fedoraproject.org and admin.fedoraproject.org/accounts to the new account system
fas was redirecting to the old https://admin.fedoraproject.org/accounts
and that in turn was still trying to proxy to the old fas2 app on the
ocp3 cluster. That app is retired and that cluster is long gone too, so
we should just redirect it to the current account system.

Signed-off-by: Kevin Fenzi <kevin@scrye.com>