Import restart_services.yml in buildvm-osbuild.yml playbook
The buildvm-osbuild.yml playbook fails when running the base role,
specifically when configuring chronyd and restarting it. The reason is
that the role is notifying a handler, which is not defined in the role,
but it is defined in the main handlers file restart_services.yml.

Import handlers from restart_services.yml in the buildvm-osbuild.yml
playbook to ensure that all common handlers are available.

Signed-off-by: Tomáš Hozza <thozza@redhat.com>