kojipkgs: set lookupcache=positive on nfs mounts
We are hitting a NFS caching issue with openqa (at least, possibly
others) where a fetch of repomd.xml will hit one of the kojipkgs servers
and then fetching the rest of the repodata will hit the other that still
has the previous koji buildroot repodata, resulting in 404's.

So, setting this should make nfs only cache postive lookups (ie, file
exists in dir), not negative ones (file does not exist). It will recheck
before returing a 404, which should prevent this issue.

Hopefully this theory is correct. We will see. :)

Signed-off-by: Kevin Fenzi <kevin@scrye.com>