Grant sysadmin-osbuild shell and sudo access to osbuild workers
The IPA client groups configuration for osbuild workers was missed in
the original PR#1498 [1]. As a result, no member of the
`sysadmin-osbuild` FAS group can SSH to the osbuild workers.

Set the appropriate IPA client variables to grant access for this
group.

Also grant access for the `sysadmin-releng` group. There is no specific
reason, it just felt sensible since osbuild workers are running in the
Fedora infrastructure.

[1] https://pagure.io/fedora-infra/ansible/pull-request/1498

Signed-off-by: Tomáš Hozza <thozza@redhat.com>