buildhw: login to docker registry
This is needed to get the new flatpak building setup working entirely.
We already added this to buildvm's, but missed adding it to buildhw
playbook at the same time. It shouldn't cause any problems for anything,
it's just a podman login call so the flatpak koji plugin can push the
completed container to the candidate registry.

Signed-off-by: Kevin Fenzi <kevin@scrye.com>