koji_hub: send flatpak builds to the flatpak channel
Right now flatpak builds use the 'default' channel, which is most
builders. Unfortunately, they get s390x builders from time to time and
those builders are in another datacenter that cannot talk to the
candidate-registry to upload resulting containers.

So, to work around this, we will setup a 'flatpak' channel that
just has machines in iad2 and then this PR configures the hub to send
all such builds to those machines (avoiding the s390x ones).

Signed-off-by: Kevin Fenzi <kevin@scrye.com>