koji_hub: container policy has a typo in it
We have 'containers' allowed, but the name space is 'container'
This results in containers failing the scm policy and not being allowed
to build. We were going to do some toolbox container updates for
f37/f38, so it would be good to fix this. I can't see it affecting
anything else really.

Signed-off-by: Kevin Fenzi <kevin@scrye.com>