bodhi: switch skopeo command
bodhi was using a small script to copy flatpak images from candidate
registry to registry. This was done because skopeo wasn't able to copy
multi-arch containers at the time. It has since gained the ability to do
that and we should stop using the temporary script to do this.

The script doesn't understand the new f39+ flatpaks hash passed to it so
it fails to push them.

This patch switches bodhi to use normal skopeo and pass it -a which
should tell it to copy all the arches, not just the one it's running on.

Signed-off-by: Kevin Fenzi <kevin@scrye.com>