proxies: registry: let ppc64le builders directly access registry
f39 adds flatpaks for ppc64le, so we need to allow ppc64le builders to
access the registry directly so they can install flatpaks in the ostree
install images. Without this they try and get them from the cdn and the
builder firewall blocks them and it times out and the image fails to
compose.

Signed-off-by: Kevin Fenzi <kevin@scrye.com>