openshift: validate when applying objects
Right now if there's a yml thats indented incorrectly, but parts of it
are fine, oc apply will apply those parts that are valid and just ignore
the rest. So, lets add a --validate=strict to have it reject these
entirely, so we know they have something wrong in them and we can fix
it, rather than being confused that they didn't apply anything.

Signed-off-by: Kevin Fenzi kevin@scrye.com