db-koji01: more tuning
Tweak some of the knobs to try and get things more happy with the heavy
load it's under. Note that cpus change will need a reboot before it's
active. The other changes have all been applied and it seems to be doing
a little better with this. Perhaps with some more cpus we can get it
stable (but yet not swamping i/o).

Signed-off-by: Kevin Fenzi <kevin@scrye.com>