postgresql: tune db01 to handle higher loads
We have been running on the defaults on db01 for a while here, but bodhi
seems to be having db query speed issues, so lets try and beef up the
postgresql config.

Increase shared_buffers as we have a lot of memory.
Increase workers as we have 24 cpus and 8 workers (default) might be a
bottleneck. Increase the related worker settings with base workers.

Signed-off-by: Kevin Fenzi <kevin@scrye.com>