copr: experiment with mz2-2x16 instances in IBM Cloud
These are about 1.5x more expensive, so we'll have only 2/3 of them for
now.  The reason for this experiment is that the old instances cz2-2x4
are often useless for memory-intensive tasks that simply do too much
swapping - so we too much time on them (which is even often wasted for
unnecessary failure).

Relates: https://github.com/fedora-copr/copr/issues/3060