openqa/worker: install ffmpeg-free
I'm adding this as a Recommends: for os-autoinst, but want to
get it on the workers now. Having it installed gives us better
videos of test runs (the internal video encoder is a bit wonky
and produces videos that have errors which make jumping around
within the video not work properly).

Signed-off-by: Adam Williamson <awilliam@redhat.com>