fmn: bump alert levels way way up
FMN sometimes gets a lot of messages in it's backlog when it's
refreshing it's cache or when there's just a ton of new messages. Since
we don't do anything except be annoyed at the alert, lets bump these way
up to alert less.

Really the case we want to alert on is if there's no consumer or if it
stays high for a long time without making any headway.

Signed-off-by: Kevin Fenzi <kevin@scrye.com>