openqa-a64-worker04: cut worker count a little
We seem to be getting lots of failures on compose tests, I think
because of how many tests kick off all at once, because they
seem to work OK when restarted one by one. Maybe this worker
count is a bit too ambitious (though it used to work), let's
try cutting it a bit.

Signed-off-by: Adam Williamson <awilliam@redhat.com>