buildvm-a64-01 to 10: try moving to new machine and f39
Lets move 10 of these over to one of the new machines and switch them to
f39. Note that we have a createrepo_c in f39-infra repo with a Epoch: 1
so it should keep using the older version instead of 1.0.x which will
cause problems for epel7 builds.

Also, add more cpus as the new machines have more cpus.
Hopefully this won't cause memory issues, if so we can back them down.

Signed-off-by: Kevin Fenzi <kevin@scrye.com>