robosignatory: resign f40 tagged packages with f41 key
This is to prep for branching next week.
We want to set this so we can tell robosgnatory to resign all the f40
tagged packages with f41. Also, newly tagged packages will also be
signed, so by the time we branch everything should be signed for both
branched (f40) and new rawhide (f41).

Signed-off-by: Kevin Fenzi <kevin@scrye.com>