nagios / koji: drop ssl cert check
This check was from long ago when koji used a self signed cert/ca
It still amusingly has that configured, so this check is telling us that
that self signed cert that we dont use anymore is expiring. :)
So, just drop this, koji is being proxies now and uses our main wildcard
cert.

Signed-off-by: Kevin Fenzi <kevin@scrye.com>